+++ docs/changes.md
@@ -1,6 +1,9 @@
 # dev
 Release date: `2018-??-??`
 
+### Core
+- [NXDRIVE-1085](https://jira.nuxeo.com/browse/NXDRIVE-1085): Review the Auto-Lock feature
+
 ### Tests
 - [NXDRIVE-887](https://jira.nuxeo.com/browse/NXDRIVE-887): Integrate SonarCloud code quality check